Fix (int) $this->session->userdata('iduser'); was (float) $this->session->userdata('iduser');

+++ application/modules/genias/controllers/genias.php
@@ -28,7 +28,7 @@ class Genias extends MX_Controller {
         $this->module_url = base_url() . 'genias/';
         //----LOAD LANGUAGE
         $this->lang->load('library', $this->config->item('language'));
-        $this->idu = (float) $this->session->userdata('iduser');
+        $this->idu = (int) $this->session->userdata('iduser');
 
         ini_set('xdebug.var_display_max_depth', 100);
